+++ fpga/tang-138k/console_top.sv
@@ -13,7 +13,7 @@ module top (
     .rst_ni (rst_ni)
   );
 
-  vv_dual_bringup #(
+  vv_cluster_bringup #(
     .CLK_HZ (50_000_000),
     .UART_BAUD (115_200)
   ) u_soc (
@@ -23,10 +23,15 @@ module top (
     .uart_tx_o (UART_TXD),
     .led_o (led),
     .debug_mailbox32_o (),
-    .debug_mailbox64_o (),
+    .debug_mailbox_p_o (),
+    .debug_mailbox_e_o (),
     .debug_halted32_o (),
-    .debug_halted64_o (),
+    .debug_halted_p_o (),
+    .debug_halted_e_o (),
     .debug_cause32_o (),
-    .debug_cause64_o ()
+    .debug_cause_p_o (),
+    .debug_cause_e_o (),
+    .debug_p_released_o (),
+    .debug_e_released_o ()
   );
 endmodule